feat: add confirmation dialog for logout action in side navigation

+++ src/app/shared/ui/layout/side-nav/side-nav.component.ts
@@ -6,6 +6,7 @@ import { Router, RouterModule } from '@angular/router';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatButtonModule } from '@angular/material/button';
 import { CookieService } from 'ngx-cookie-service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-side-nav',
@@ -24,8 +25,20 @@ export class SideNavComponent {
   router = inject(Router);
   cookieService = inject(CookieService);
   logout(): void {
-    this.cookieService.delete('access_token');
-    this.cookieService.delete('refresh_token');
-    this.router.navigate(['/login']);
+    Swal.fire({
+      icon: 'question',
+      title: 'Cerrar Sesión',
+      text: 'Estas seguro de cerrar sesión?',
+      showCancelButton: true,
+      confirmButtonText: 'Cerrar',
+      cancelButtonText: 'Cancelar',
+      confirmButtonColor: '#3085d6',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.cookieService.delete('access_token');
+        this.cookieService.delete('refresh_token');
+        this.router.navigate(['/auth']);
+      }
+    })
   }
 }
